easy peasy, added an endpoint for deleting, a fetch call for deleting, and added a delete button with a modal on the client side
